bugfix: incorrect parsing of Long using Integer.parseInt() instead of Long.parseLong()

+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/methodreplacement/classes/LongClassReplacement.java
@@ -29,7 +29,7 @@ public class LongClassReplacement implements MethodReplacementClass {
         }
 
         if (idTemplate == null) {
-            return Integer.parseInt(input);
+            return Long.parseLong(input);
         }
 
         try {
